refactor: split commands in multiple files
Context:
  * one big main.rs file with code for all commands

Problem:
  * hard to read
  * not easy to add unit test for each commands

Solution:
  * split main.rs file into multiple files with one file per command
  * only copy/paste commands in his own file with good imports, so no
    regression should happen
